backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
pydantic==2.6.1
python-multipart==0.0.9
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
joblib==1.3.2
